+++ src/main/java/es/upm/pproject/geditor/model/SVGRectangle.java
@@ -4,14 +4,10 @@ import java.awt.Color;
 import java.awt.geom.Rectangle2D;
 
 public class SVGRectangle extends SVGElement {
-    private int width;
-    private int height;
+    private double width;
+    private double height;
 
-    // public SVGRectangle() {
-
-    // }
-
-    public SVGRectangle(int x, int y, int width, int height, Color fillColor, double fillOpacity, Color strokeColor,
+    public SVGRectangle(double x, double y, double width, double height, Color fillColor, double fillOpacity, Color strokeColor,
             double strokeOpacity, double strokeWidth) {
         super(x, y, fillColor, fillOpacity, strokeColor, strokeOpacity, strokeWidth);
         this.width = width;
@@ -28,19 +24,19 @@ public class SVGRectangle extends SVGElement {
     }
 
     // Getters and setters
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 }
